+++ src/Data/TypedEncoding/Internal/Types/Common.hs
@@ -1,4 +1,4 @@
 module Data.TypedEncoding.Internal.Types.Common where
 
 -- | Represents value level (single) annotation.
-type SomeAnn = String    
+type EncAnn = String