+++ src/Examples/TypedEncoding/Overview.hs
@@ -69,14 +69,14 @@ helloB64Decoded = fromEncoding . decodeAll $ helloB64
 -- The above example start by placing payload in zero-encoded @Enc '[] ()@ type and then apply 'recreateFAll'
 -- this is a good way to recreate encoded type if encoding is known. 
 --
--- It is it not, existentially quantified 'Unchecked' type can be used. 
+-- If is it not, 'UncheckedEnc' type can be used. 
 --
 -- (See 'Examples.TypedEncoding.ToEncString' for better example).
 -- 
 -- This module is concerned only with the first approach. 
 --
--- >>> let unchecked = toUnchecked ["enc-B64"] () ("SGVsbG8gV29ybGQ=" :: T.Text)
--- >>> verifyUnchecked' @'["enc-B64"] unchecked
+-- >>> let unchecked = toUncheckedEnc ["enc-B64"] () ("SGVsbG8gV29ybGQ=" :: T.Text)
+-- >>> verifyUncheckedEnc' @'["enc-B64"] unchecked
 -- Just (Right (MkEnc Proxy () "SGVsbG8gV29ybGQ="))
 helloB64Recovered :: Either RecreateEx (Enc '["enc-B64"] () B.ByteString)
 helloB64Recovered = recreateFAll . toEncoding () $ "SGVsbG8gV29ybGQ="